+++ example/lib/soundEffect/sound_effect.dart
@@ -86,6 +86,8 @@ class _SoundEffectState extends State<SoundEffect> {
       codec: Codec.pcm16,
       numChannels: _tNumChannels,
       sampleRate: _tSampleRate,
+      bufferSize:1024,
+      interleaved: true,
     );
   }